+++ tests/ConsistentHashTest.php
@@ -18,6 +18,7 @@ namespace Esi\ConsistentHash\Tests;
 use Esi\ConsistentHash\ConsistentHash;
 use Esi\ConsistentHash\Exception\TargetException;
 use Esi\ConsistentHash\Hasher\Crc32Hasher;
+use Esi\ConsistentHash\Hasher\Xxh32Hasher;
 use Esi\ConsistentHash\Tests\Hasher\MockHasher;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\UsesClass;
@@ -31,6 +32,7 @@ use function range;
 #[CoversClass(ConsistentHash::class)]
 #[CoversClass(TargetException::class)]
 #[UsesClass(Crc32Hasher::class)]
+#[UsesClass(Xxh32Hasher::class)]
 class ConsistentHashTest extends TestCase
 {
     public function testAddTargetAndGetAllTargets(): void